Add files via upload
Tools and libraries that can help with this project include Python, scikit-learn, Pandas, NumPy, and more advanced libraries like XGBoost or TensorFlow for complex models. Additionally, you might want to consider using Jupyter notebooks for data exploration and visualization.

Remember that the success of your house price prediction model depends on the quality and quantity of data, as well as the choice of the right algorithm and careful data preprocessing. Additionally, interpretability and fairness in housing prediction models are